move all non-pages out of `/pages`
