+++ containers/mining/apps/table-section.js
