+++ containers/mining/faq/faq-section.js
@@ -1,9 +1,9 @@
 import * as React from 'react'
-import { Heading, Section } from '@pages/mining/shared'
+import { Heading, Section } from '@containers/mining/shared'
 import { Box } from '@components/mining'
 import styled from 'styled-components'
 
-import Content from './faq-content.md'
+import Content from '@pages/mining/faq/faq-content.md'
 
 export const FAQwrapper = styled(Box)`
   & h2 {
