+++ containers/mining/sections/faq/index.js
