+++ containers/mining/sections/hero/index.js
