+++ containers/mining/sections/how-much-earn/index.js
