+++ containers/mining/sections/pioneers/index.js
