+++ containers/mining/sections/ready-apps/index.js
