+++ containers/mining/sections/start-app-mining/index.js
