+++ containers/mining/shared.js
