+++ containers/mining/starter-kit-modal.js
