+++ pages/mining/apps/index.js
@@ -3,7 +3,7 @@ import Head from '@containers/head'
 import { MiningPage } from '@components/mining/page'
 import { selectApiServer, selectAppMiningApps } from '@stores/apps/selectors'
 import { connect } from 'react-redux'
-import { AllApps } from '@pages/mining/apps/table-section'
+import { AllApps } from '@containers/mining/apps/table-section'
 import { ModalRoot } from 'blockstack-ui'
 import { Header } from '@components/mining/header'
 import { Footer } from '@components/mining/footer'
