+++ pages/mining/index.js
@@ -3,12 +3,12 @@ import Head from '@containers/head'
 import { MiningPage } from '@components/mining/page'
 import { selectApiServer, selectAppMiningMonths, selectAppMiningApps } from '@stores/apps/selectors'
 import { connect } from 'react-redux'
-import { StartAppMiningSection } from '@pages/mining/sections/start-app-mining'
-import { Hero } from '@pages/mining/sections/hero'
-import { HowMuchSection } from '@pages/mining/sections/how-much-earn'
-import { RankingSection } from '@pages/mining/sections/ranking'
-import { PioneersSection } from '@pages/mining/sections/pioneers'
-import { FAQSection } from '@pages/mining/sections/faq'
+import { StartAppMiningSection } from '@containers/mining/sections/start-app-mining'
+import { Hero } from '@containers/mining/sections/hero'
+import { HowMuchSection } from '@containers/mining/sections/how-much-earn'
+import { RankingSection } from '@containers/mining/sections/ranking'
+import { PioneersSection } from '@containers/mining/sections/pioneers'
+import { FAQSection } from '@containers/mining/sections/faq'
 import { ModalRoot } from 'blockstack-ui'
 import { Header } from '@components/mining/header'
 import { Footer } from '@components/mining/footer'
